add plot config file
